Creaye and make game
We made a system where the Unity host makes a game room and gets a code, and people on the website can join only if they have that code. The server keeps track of who’s in the room and tells the host so they can see all the players.